Interim commit with Racket as the server.
Committing as a checkpoint before experimenting with flipping this and
making Emacs the server. (Why? After we `make-comint' there is of course
some delay while Racket and our program load -- during which
`open-network-stream' will fail to make a connection because the TCP
server is now yet running. So, it might work more smoothly for Emacs to
be the "server" and wait for the Racket backend to connect. After the
connection, "server" and "client" is N/A -- Emacs will still be issuing
commands, and the Racket backend responds, as before.)