Emacs listens for the connection.
Why not have Racket, which is really "the server", listen for the
connection? After we `make-comint' there is of course some delay while
Racket and our Racket backend program load. During which, an attempt to
`open-network-stream' to make a connection would fail, because the
Racket TCP server wouldn't be running, yet.

Instead it's smoother to have Emacs listen, and wait for the Racket
backend to connect when it's ready.

Note that Emacs is the "server" only in the sense that it listens for
the initial connection to be made. Thereafter, using that connection,
Emacs issues commands, and the Racket backend responds; in that sense
Emacs is the "client" and the Racket backend is the "server".