Support submodules
,run accepts '(submod file.rkt mod ...+)

racket-run runs the submodule enclosing point.

racket-test is now ,run `(submod ,buffer-file-name test) as opposed to
squirting a require into the REPL. As a result, the user is in the test
submodule and can experiment with tests. (Caveat: See bug note below.)

The hairiest part is handling the variations of quoted vs. unquoted
paths, paths vs. module-paths, absolute vs. relative paths, etc. A new
`mod` struct and related functions manage this.

There is a bug in Racket 6.2.1 and earlier (but fixed on HEAD) that,
when errortrace is being used, module->namespace does not return the
definitions in a module+ form. The module is evaluated fine (e.g. tests
in a test submodule are run), but the definitions cannot be accessed in
the REPL. (Plain `module` and non-#f `module*` forms do work fine in
6.2.1 and earlier.) So, display warning/advice the first time a user
does this each session.