check-syntax: motion and tooltips
- Use cursor-sensor-mode when available (instead of point motion
  hooks).

- Use tooltips not message

This started off from reviewing PR #242 to fix #243,

First, it wasn't working for me on Emacs 25.

After I got that working, I realized we could display as tooltips
instead of in the echo area. Having the information in proximity is
helpful. (If it were to turn out that tooltips per se are a bad idea,
then a text overlay could be used instead.)